Fix switching to longer queue near the end of job's lifetime
The code assumed that the LSF time limits are based on the CPU
time. However, it turns out that the limits are normalized by the
execution host's CPU factor and also depends on the settings.

Absolute wall-clock run time seems to be the safest (and easiest)
bet, and works well in practice.