An attempt to fix rare failures to increase memory limits
Jobs occasionally fail to recognise the increased memory limits. This may be
because the failed status is visible to the spawner earlier than the
increase-request file created on the remote file system. This small timeout
should help without causing any harm - the job which requested the increase is
just going to sit idle for one minute before exiting with an error status (see
run-eagle for an example)