Fix updates of .r files.
The introduction of time stamp (b6a4b12b63) caused that the .r files had
to be always updated. (Also it was incorrect in that it would not show
life time, but the start time of the script.)

The consequent fix 790873942 was neglecting changes made by editing the user
scripts.

This commit addresses both issues.